# src/backend/docs/swagger.yaml
basePath: /api
definitions:
  dto.LoginInput:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: strongpassword
        type: string
    required:
    - email
    - password
    type: object
  dto.MessageResponse:
    properties:
      message:
        example: ลงทะเบียนสำเร็จ
        type: string
    type: object
  dto.RegisterInput:
    properties:
      email:
        example: john@example.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: strongpassword
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  dto.UserProfileResponse:
    properties:
      email:
        example: john@example.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: John Doe
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Backend API for managing room reservation system.
  title: Room Reservation API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login Input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: Login Success
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      summary: Login
      tags:
      - Auth
  /auth/me:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: User Profile
          schema:
            $ref: '#/definitions/dto.UserProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Get User Profile
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: สร้างบัญชีใหม่ในระบบ
      parameters:
      - description: ข้อมูลผู้ใช้ใหม่
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      summary: ลงทะเบียนผู้ใช้
      tags:
      - Auth
securityDefinitions:
  BearerAuth:
    description: Enter JWT Bearer token **_only_**
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"